Refactor OCR and Firebase integration; remove legacy database code and implement EasyOCR and Firestore for card verification
